requirements.txt
django
psycopg2-binary
python-docx
PyPDF2
pandas
spacy
python-dotenv